Rename page to account with new path, add vote proxy section
